Fixed compiler warnings

+++ Neighbor.h
@@ -44,7 +44,7 @@ public:
 			int neighbors[2];
 			Upward::cells(puml, puml.faces()[faces[i]], neighbors);
 
-			if (neighbors[0] == clid)
+			if (static_cast<unsigned int>(neighbors[0]) == clid)
 				flid[i] = neighbors[1];
 			else
 				flid[i] = neighbors[0];
